move common fns to utils so torch is optional
